rasa-3.6.x/my_rule_bot/requirements.txt
rasa==3.6.20
rasa-sdk==3.6.2
numpy==1.23.5
packaging==20.9
tensorflow==2.11.0
protobuf==4.23.3
spacy==3.5.3
scipy==1.10.1
scikit-learn==1.2.2
SQLAlchemy==1.4.46
redis==4.5.5
aiohttp==3.8.5
pydantic==1.10.9
PyJWT==2.8.0
psycopg2-binary